Moves ConfigProvider to subnamespace
In order to prevent conflicts with other router implementations, moves
the `ConfigProvider` to a new subnamespace,
`Zend\Expressive\Router\FastRouteRouter\ConfigProvider`.
